Fix regression with Nix 2.19
Apparently newer Nix versions are again displaying store paths in the
format

  `/nix/store/<hash>-x`

when running `nix path-info --all --json`.

This caused `cache-gc` to delete a lot of paths to delete that shouldn't
be deleted leaving my Hydra with a corrupted binary cache.

While this certainly needs an upstream bugreport, I decided to hack
around that by reading the `.narinfo` files directly to get the correct
store path because I wanted a quick fix.

I originally wanted to do that with a second slurpfile, but the IO
involved with reading >10k `.narinfo` files caused a significant
slowdown on my end, so I decided to rewrite the `jq`-part with Python so
that I never have to shell out for the store-path detection or
performing `stat(1)`.